refactor(core/common): split utils.go and nameidmapper.go into internal packages
Retires the core/common grab-bag: cache-path/key construction moves to
internal/cachekey and target-graph wire-format conversion (ID mapping +
chunking) moves to internal/targetgraph, matching the existing
internal/targetdiff convention for implementation-detail helpers shared
across controller/orchestrator. core/common now only holds error
classification (errors.go).
