+++ controller/BUILD.bazel
@@ -17,6 +17,8 @@ go_library(
         "//config",
         "//core/common",
         "//core/storage",
+        "//internal/cachekey",
+        "//internal/targetgraph",
         "//orchestrator",
         "//tangopb",
         "@com_github_uber_go_tally//:tally",
@@ -40,6 +42,7 @@ go_test(
         "//core/common",
         "//core/storage",
         "//core/storage/storagemock",
+        "//internal/targetgraph",
         "//orchestrator/orchestratormock",
         "//tangopb",
         "//tangopb/tangopbmock",
