+++ example/cmd/query-bench/BUILD.bazel
@@ -7,8 +7,8 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//core/bazel",
-        "//core/common",
         "//core/targethasher",
+        "//internal/targetgraph",
         "@com_github_gogo_protobuf//jsonpb",
         "@org_uber_go_zap//:zap",
     ],
