+++ orchestrator/BUILD.bazel
@@ -18,6 +18,8 @@ go_library(
         "//core/storage",
         "//core/workspace",
         "//graphrunner",
+        "//internal/cachekey",
+        "//internal/targetgraph",
         "//tangopb",
         "@com_github_uber_go_tally//:tally",
         "@org_uber_go_zap//:zap",
